Made corrections and additions found necessary while doing network discovery in Saratoga-StackZ Corrected returned source address being 32 bits instead of 64. Sending the rest of the Beacon Indication Parameters instead of just the Pan Descriptors during an active scan.

+++ include/net/nl802154.h
@@ -320,7 +320,7 @@ struct ieee802154_beacon_indication {
 	struct {
 		u8 src_addr_mode;          /* enumeration:  SHORT_ADDR, EXTENDED_ADDR   */
 		u16 src_pan_id;            /* Integer    :  0x0000 -- 0xffff            */
-		u32 src_addr;              /* type?      :  range ?                     */
+		u64 src_addr;              /* type?      :  range ?                     */
 		u8 channel_num;            /* integer    :  11 - 27?                    */
 		u8 channel_page;           /* integer    :  range ?                     */
 		u8 superframe_spec;        /* bitfield   :                              */
